Serialization fix
