+++ SeniorApp/app/src/main/java/com/example/seniorapp/AppInfo.kt
@@ -1,5 +1,29 @@
 package com.example.seniorapp
 
-import android.graphics.drawable.Drawable
+import android.os.Parcel
+import android.os.Parcelable
 
-data class AppInfo(val packageName: String, val label: String, val icon: Drawable)
+data class AppInfo(val packageName: String, val label: String
+) : Parcelable {
+    constructor(parcel: Parcel) : this(
+        parcel.readString() ?: "",
+        parcel.readString() ?: ""
+    )
+
+    override fun writeToParcel(parcel: Parcel, flags: Int) {
+        parcel.writeString(packageName)
+        parcel.writeString(label)
+    }
+
+    override fun describeContents(): Int = 0
+
+    companion object CREATOR : Parcelable.Creator<AppInfo> {
+        override fun createFromParcel(parcel: Parcel): AppInfo {
+            return AppInfo(parcel)
+        }
+
+        override fun newArray(size: Int): Array<AppInfo?> {
+            return arrayOfNulls(size)
+        }
+    }
+}
