+++ SeniorApp/app/src/main/java/com/example/seniorapp/Navigation.kt
@@ -19,17 +19,17 @@ fun AppNavHost(navController: NavHostController) {
         composable(
             "home",
             enterTransition = {
-                // Fade in and slide in from bottom
-                fadeIn(tween(800)) + slideInHorizontally(
-                    initialOffsetX = { 1000 }, // Start from the right
-                    animationSpec = tween(800)
+                // Fade in with a smooth slide-in from the right
+                fadeIn(tween(500)) + slideInHorizontally(
+                    initialOffsetX = { 500 }, // Slide from the right
+                    animationSpec = tween(500)
                 )
             },
             exitTransition = {
-                // Fade out and slide out to top
-                fadeOut(tween(800)) + slideOutHorizontally(
-                    targetOffsetX = { -1000 }, // Slide out to the left
-                    animationSpec = tween(800)
+                // Fade out with a smooth slide-out to the left
+                fadeOut(tween(500)) + slideOutHorizontally(
+                    targetOffsetX = { -500 }, // Slide to the left
+                    animationSpec = tween(500)
                 )
             }
         ) {
@@ -40,17 +40,17 @@ fun AppNavHost(navController: NavHostController) {
         composable(
             "settings",
             enterTransition = {
-                // Fade in and slide in from bottom
-                fadeIn(tween(800)) + slideInHorizontally(
-                    initialOffsetX = { 1000 }, // Start from the right
-                    animationSpec = tween(800)
+                // Fade in with a smooth slide-in from the right
+                fadeIn(tween(500)) + slideInHorizontally(
+                    initialOffsetX = { 500 }, // Slide from the right
+                    animationSpec = tween(500)
                 )
             },
             exitTransition = {
-                // Fade out and slide out to top
-                fadeOut(tween(800)) + slideOutHorizontally(
-                    targetOffsetX = { -1000 }, // Slide out to the left
-                    animationSpec = tween(800)
+                // Fade out with a smooth slide-out to the left
+                fadeOut(tween(500)) + slideOutHorizontally(
+                    targetOffsetX = { -500 }, // Slide to the left
+                    animationSpec = tween(500)
                 )
             }
         ) {
